Move serializers in `Types` and improve `.extractCertification`
